third_party/blink/web_tests/fast/backgrounds/background-position-parsing-2-expected.txt

Test to make sure background-position is parsed correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

background-position with one value
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is '70% center'
PASS computedStyle.backgroundPosition is '70% 50%'
PASS style.backgroundPosition is '84px center'
PASS computedStyle.backgroundPosition is '84px 50%'
PASS style.backgroundPosition is 'left center'
PASS computedStyle.backgroundPosition is '0% 50%'
PASS style.backgroundPosition is 'right center'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is 'center bottom'
PASS computedStyle.backgroundPosition is '50% 100%'
PASS style.backgroundPosition is 'center top'
PASS computedStyle.backgroundPosition is '50% 0%'
PASS style.backgroundPosition is 'center center'
PASS computedStyle.backgroundPosition is '50% 50%'
background-position with two values
PASS style.backgroundPosition is 'left bottom'
PASS computedStyle.backgroundPosition is '0% 100%'
PASS style.backgroundPosition is 'left bottom'
PASS computedStyle.backgroundPosition is '0% 100%'
PASS style.backgroundPosition is '100% bottom'
PASS computedStyle.backgroundPosition is '100% 100%'
PASS style.backgroundPosition is '100% top'
PASS computedStyle.backgroundPosition is '100% 0%'
PASS style.backgroundPosition is '54px bottom'
PASS computedStyle.backgroundPosition is '54px 100%'
PASS style.backgroundPosition is 'center center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '5% bottom'
PASS computedStyle.backgroundPosition is '5% 100%'
PASS style.backgroundPosition is '30pt -20px'
PASS computedStyle.backgroundPosition is '40px -20px'
PASS style.backgroundPosition is 'right center'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is '100% 0px'
PASS computedStyle.backgroundPosition is '100% 0px'
PASS style.backgroundPosition is 'right center'
PASS computedStyle.backgroundPosition is '100% 50%'
PASS style.backgroundPosition is 'center 50%'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is 'left center'
PASS computedStyle.backgroundPosition is '0% 50%'
PASS style.backgroundPosition is '-20% center'
PASS computedStyle.backgroundPosition is '-20% 50%'
PASS style.backgroundPosition is 'right top'
PASS computedStyle.backgroundPosition is '100% 0%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
background-position invalid with one or two values, no change expected
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '50% center'
PASS computedStyle.backgroundPosition is '50% 50%'
PASS style.backgroundPosition is '1px 1px'
PASS computedStyle.backgroundPosition is '1px 1px'
background-position with CSS3 comma separator, one or two values
PASS style.backgroundPosition is '50% center, 100% center'
PASS computedStyle.backgroundPosition is '50% 50%, 100% 50%'
PASS style.backgroundPosition is 'center top, center bottom'
PASS computedStyle.backgroundPosition is '50% 0%, 50% 100%'
PASS style.backgroundPosition is 'right center, 5px bottom'
PASS computedStyle.backgroundPosition is '100% 50%, 5px 100%'
PASS style.backgroundPosition is 'left top, right bottom'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
background-position with CSS3 comma separator, with invalid one or two values, no change expected
PASS style.backgroundPosition is 'left top, right bottom'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
PASS style.backgroundPosition is 'left top, right bottom'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
PASS style.backgroundPosition is 'left top, right bottom'
PASS computedStyle.backgroundPosition is '0% 0%, 100% 100%'
background-position with CSS3 four values
PASS style.backgroundPosition is 'left 10px top 15px'
PASS computedStyle.backgroundPosition is 'left 10px top 15px'
PASS style.backgroundPosition is 'left 10% top 30%'
PASS computedStyle.backgroundPosition is 'left 10% top 30%'
PASS style.backgroundPosition is 'left 30% top 10%'
PASS computedStyle.backgroundPosition is 'left 30% top 10%'
PASS style.backgroundPosition is 'right top 15px'
PASS computedStyle.backgroundPosition is '100% top 15px'
PASS style.backgroundPosition is 'left 10px center'
PASS computedStyle.backgroundPosition is 'left 10px 50%'
PASS style.backgroundPosition is 'center top 20px'
PASS computedStyle.backgroundPosition is '50% top 20px'
PASS style.backgroundPosition is 'center top 20px'
PASS computedStyle.backgroundPosition is '50% top 20px'
PASS style.backgroundPosition is 'left 30px center'
PASS computedStyle.backgroundPosition is 'left 30px 50%'
PASS style.backgroundPosition is 'left 20% top'
PASS computedStyle.backgroundPosition is 'left 20% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
background-position with CSS3 invalid four values, no change expected
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
PASS style.backgroundPosition is ''
PASS computedStyle.backgroundPosition is '0% 0%'
background-position with CSS3 four values and comma
PASS style.backgroundPosition is 'center center, left bottom 20px'
PASS computedStyle.backgroundPosition is '50% 50%, 0% bottom 20px'
PASS style.backgroundPosition is 'right 20px bottom 20px, left center'
PASS computedStyle.backgroundPosition is 'right 20px bottom 20px, 0% 50%'
PASS style.backgroundPosition is 'left 10px top 15px, right 20% bottom 20px'
PASS computedStyle.backgroundPosition is 'left 10px top 15px, right 20% bottom 20px'
PASS style.backgroundPosition is 'left 10% top, right top 0px'
PASS computedStyle.backgroundPosition is 'left 10% 0%, 100% top 0px'
PASS style.backgroundPosition is 'right top 15px, right 20px bottom'
PASS computedStyle.backgroundPosition is '100% top 15px, right 20px 100%'
background-position with invalid CSS3 four values and comma, no change expected
PASS style.backgroundPosition is 'right top 15px, right 20px bottom'
PASS computedStyle.backgroundPosition is '100% top 15px, right 20px 100%'
PASS style.backgroundPosition is 'right top 15px, right 20px bottom'
PASS computedStyle.backgroundPosition is '100% top 15px, right 20px 100%'
PASS style.backgroundPosition is 'right top 15px, right 20px bottom'
PASS computedStyle.backgroundPosition is '100% top 15px, right 20px 100%'
background-position inside the background shorthand
PASS style.background is 'left 30% top 10% / 10em round fixed border-box border-box gray'
PASS computedStyle.background is 'rgb(128, 128, 128) none round fixed left 30% top 10% / 160px border-box border-box'
PASS style.background is 'right top 15px / 10em round fixed border-box border-box gray'
PASS computedStyle.background is 'rgb(128, 128, 128) none round fixed 100% top 15px / 160px border-box border-box'
PASS style.background is 'left 10px center / 10em round fixed border-box border-box gray'
PASS computedStyle.background is 'rgb(128, 128, 128) none round fixed left 10px 50% / 160px border-box border-box'
PASS style.background is 'left 10px center round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed left 10px 50% / auto border-box border-box'
PASS style.background is 'center top 20px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed 50% top 20px / auto border-box border-box'
PASS style.background is 'center top 20px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed 50% top 20px / auto border-box border-box'
PASS style.background is 'center top round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed 50% 0% / auto border-box border-box'
PASS style.background is '50px 60px / 50px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed 50px 60px / 50px border-box border-box'
PASS style.background is '50px center / 50px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed 50px 50% / 50px border-box border-box'
PASS style.background is 'left top 60px / 50px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed 0% top 60px / 50px border-box border-box'
PASS style.background is 'left -20px top 60px / 50px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed left -20px top 60px / 50px border-box border-box'
PASS style.background is 'left 20px top / 50px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed left 20px 0% / 50px border-box border-box'
PASS style.background is 'left 20px top / 50px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed left 20px 0% / 50px border-box border-box'
PASS style.background is 'left 20px top 40px / 50px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed left 20px top 40px / 50px border-box border-box'
PASS style.background is 'left 20px top 40px round fixed border-box border-box'
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none round fixed left 20px top 40px / auto border-box border-box'
PASS style.background is 'left 30% top 10% round fixed border-box border-box gray'
PASS computedStyle.background is 'rgb(128, 128, 128) none round fixed left 30% top 10% / auto border-box border-box'
background-position inside the background shorthand, some invalid cases, no changes expected
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS style.background is ''
PASS computedStyle.background is 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box'
PASS successfullyParsed is true

TEST COMPLETE
